Moved geopyspark.geotrellis.constants to geopyspark.constants
Fixed typo
