+++ geopyspark/tests/geotiff_raster_rdd_test.py
@@ -2,10 +2,10 @@ import unittest
 from os import walk, path
 import rasterio
 
+from geopyspark.constants import SPATIAL
 from geopyspark.tests.python_test_utils import check_directory, geotiff_test_path
 from geopyspark.geotrellis.geotiff_rdd import geotiff_rdd, geotiff_raster_rdd
 from geopyspark.tests.base_test_class import BaseTestClass
-from geopyspark.geotrellis.constants import SPATIAL
 
 
 check_directory()
