+++ geopyspark/tests/tile_layer_methods_test.py
@@ -2,6 +2,7 @@ import os
 import unittest
 import rasterio
 
+from geopyspark.constants import SPATIAL
 from geopyspark.tests.python_test_utils import check_directory, geotiff_test_path
 from geopyspark.geotrellis.tile_layer import (collect_metadata,
                                               cut_tiles,
@@ -9,7 +10,6 @@ from geopyspark.geotrellis.tile_layer import (collect_metadata,
                                               tile_to_layout)
 from geopyspark.geotrellis.geotiff_rdd import geotiff_rdd
 from geopyspark.tests.base_test_class import BaseTestClass
-from geopyspark.geotrellis.constants import SPATIAL
 
 
 check_directory()
